Fix turning after reaching crate
I think I lost track of code that I was working on because auto-formatting made it look too cromulent.
And I decided not to bother squashing commits, and so I didn't make an easy diff for me to look at.

Broken in 5527651c9e245abba025ddda890ff8d1728ec63a