Fix save slot logic mostly
When you undo a title change, I'm not sure how it should work. Maybe it should ask you what you want to do?
I could retcon history / make it not update the title when undoing/redoing.
If I moved the slot (deleting the old) instead of copying (leaving the old), I could avoid this in the most common case, but it could still happen.