Fix test expectation (Crate Fall Onto Offset Blocks)
I think the original idea was you would walk down the shallow steps and get the bin if the crate fell through the steps, so a win would be a fail, but that's not what I ended up with. I do try to align win conditions with passes. I ended up with the crate forming a bridge. It doesn't test all the nuances, like stopping falling at the highest up block, but it's alright.