Fix dinosaur test (Shallow Steps)
The teeth are supposed to be embedded in the mouth in this level.
I the Junkbot games, there was no gravity on regular bricks, but I have gravity on them, and I'm just special casing it so they can get purposely stuck in other bricks in this test case level.
This fixes the test because the teeth fell and blocked your way.