DRY: Apparently I don't care to refine the test timeouts
They can run at super speed, so it doesn't matter much.
The ones that draw (don't win or lose) aren't as fun to watch, but refining it without risking the validity of the test seems fraught.
It could still be done though later, I do like making the tests entertaining, dramatic. Tippy Toast is the best.