Fix projection coherence, especially with Separate Objects mode
When rotating (with IJKL), things now appear as coherent objects, but I see things disappearing and reappearing, so I think Three.js is culling things, without knowledge of my custom projection. Maybe I can enable the oblique projection matrix, but I'll need to force it to ignore it in the shader. Or just disable culling, which would be more surefire (I'm not totally sure the culling works with an oblique projection.)